# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Depanini Api",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/clients/{s_id}": {
            "get": {
                "tags": [
                    "clients"
                ],
                "summary": "get client details using secret id",
                "description": "get client details using secret id",
                "operationId": "get client details",
                "parameters": [
                    {
                        "name": "s_id",
                        "in": "path",
                        "description": "Client secret id ",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client details fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "no client found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/clients/register": {
            "post": {
                "tags": [
                    "clients"
                ],
                "summary": "register new client ",
                "operationId": "register_client",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "full Name"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+213664419425"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/clients/login": {
            "post": {
                "tags": [
                    "clients"
                ],
                "summary": "login a client using phone number",
                "operationId": "login_client",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+213664419425"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/clients/{s_id}/update": {
            "put": {
                "tags": [
                    "clients"
                ],
                "summary": "update a client informations",
                "operationId": "update_client",
                "parameters": [
                    {
                        "name": "s_id",
                        "in": "path",
                        "description": "Client secret id ",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "location": {
                                        "type": "object",
                                        "example": {
                                            "lang": "45558",
                                            "lat": "4587.00"
                                        }
                                    },
                                    "email": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "messaging_token": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/clients/{s_id}/update-photo": {
            "post": {
                "tags": [
                    "clients"
                ],
                "summary": "update a client photo",
                "operationId": "update_client_photo",
                "parameters": [
                    {
                        "name": "s_id",
                        "in": "path",
                        "description": "Client secret id ",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "photo": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Photo updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    }
}